Move tests and examples to compiler package
