+++ internal/compiler/examples/output1.js
