+++ internal/compiler/examples/output2.js
