+++ internal/compiler/examples/output3.js
